+++ mzLib/Test/Omics/SampleInfo/IsobaricQuantSampleInfoTests.cs
@@ -1,10 +1,10 @@
-﻿using MassSpectrometry;
-using NUnit.Framework;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
+using MassSpectrometry;
+using NUnit.Framework;
 
-namespace Test.Omics
+namespace Test.Omics.SampleInfo
 {
     /// <summary>
     /// Tests for IsobaricQuantSampleInfo class.
