+++ mzLib/Test/Omics/SampleInfo/SpectraFileInfoTests.cs
@@ -1,10 +1,9 @@
 ﻿using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
-using System.Linq;
 using MassSpectrometry;
 using NUnit.Framework;
 
-namespace Test.Omics
+namespace Test.Omics.SampleInfo
 {
     /// <summary>
     /// Tests for SpectraFileInfo class.
